Improves initial setup
